requirements.txt
Package                   Version
------------------------- --------------
annotated-types           0.7.0
anyio                     4.10.0
appnope                   0.1.4
argon2-cffi               25.1.0
argon2-cffi-bindings      25.1.0
arrow                     1.3.0
asttokens                 3.0.0
async-lru                 2.0.5
attrs                     25.3.0
babel                     2.17.0
beautifulsoup4            4.13.5
bertopic                  0.17.3
bleach                    6.2.0
cachetools                5.5.2
certifi                   2025.8.3
cffi                      1.17.1
charset-normalizer        3.4.3
click                     8.2.1
comm                      0.2.3
contourpy                 1.3.3
cycler                    0.12.1
debugpy                   1.8.16
decorator                 5.2.1
deepl                     1.22.0
defusedxml                0.7.1
executing                 2.2.0
fastjsonschema            2.21.2
filelock                  3.19.1
fonttools                 4.59.2
fqdn                      1.5.1
fsspec                    2025.7.0
google-auth               2.40.3
google-genai              1.33.0
googletrans               4.0.2
h11                       0.16.0
h2                        4.3.0
hdbscan                   0.8.40
hf-xet                    1.1.9
hpack                     4.1.0
httpcore                  1.0.9
httpx                     0.28.1
huggingface-hub           0.34.4
hyperframe                6.1.0
idna                      3.10
ipykernel                 6.30.1
ipython                   9.5.0
ipython_pygments_lexers   1.1.1
isoduration               20.11.0
jedi                      0.19.2
Jinja2                    3.1.6
joblib                    1.5.2
json5                     0.12.1
jsonpointer               3.0.0
jsonschema                4.25.1
jsonschema-specifications 2025.4.1
jupyter_client            8.6.3
jupyter_core              5.8.1
jupyter-events            0.12.0
jupyter-lsp               2.3.0
jupyter_server            2.17.0
jupyter_server_terminals  0.5.3
jupyterlab                4.4.6
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.3
kiwisolver                1.4.9
lark                      1.2.2
llvmlite                  0.44.0
MarkupSafe                3.0.2
matplotlib                3.10.6
matplotlib-inline         0.1.7
mistune                   3.1.4
mlx                       0.29.0
mlx-lm                    0.27.1
mlx-metal                 0.29.0
mpmath                    1.3.0
narwhals                  2.3.0
nbclient                  0.10.2
nbconvert                 7.16.6
nbformat                  5.10.4
nest-asyncio              1.6.0
networkx                  3.5
nltk                      3.9.1
notebook                  7.4.5
notebook_shim             0.2.4
numba                     0.61.2
numpy                     2.2.6
packaging                 25.0
pandas                    2.3.2
pandas-stubs              2.3.2.250827
pandocfilters             1.5.1
parso                     0.8.5
pexpect                   4.9.0
pillow                    11.3.0
pip                       25.1.1
platformdirs              4.4.0
plotly                    6.3.0
prettytable               3.16.0
prometheus_client         0.22.1
prompt_toolkit            3.0.52
protobuf                  6.32.0
psutil                    7.0.0
ptyprocess                0.7.0
pure_eval                 0.2.3
pyasn1                    0.6.1
pyasn1_modules            0.4.2
pycparser                 2.22
pydantic                  2.11.7
pydantic_core             2.33.2
Pygments                  2.19.2
pynndescent               0.5.13
pyparsing                 3.2.3
python-dateutil           2.9.0.post0
python-dotenv             1.1.1
python-json-logger        3.3.0
pytz                      2025.2
PyYAML                    6.0.2
pyzmq                     27.0.2
referencing               0.36.2
regex                     2025.7.34
requests                  2.32.5
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rfc3987-syntax            1.1.0
rpds-py                   0.27.1
rsa                       4.9.1
safetensors               0.6.2
scikit-learn              1.7.1
scipy                     1.16.1
Send2Trash                1.8.3
sentence-transformers     5.1.0
setuptools                80.9.0
six                       1.17.0
sniffio                   1.3.1
soupsieve                 2.8
stack-data                0.6.3
sympy                     1.14.0
tenacity                  9.1.2
terminado                 0.18.1
threadpoolctl             3.6.0
tinycss2                  1.4.0
tokenizers                0.21.4
torch                     2.8.0
torchgen                  0.0.1
tornado                   6.5.2
tqdm                      4.67.1
traitlets                 5.14.3
transformers              4.55.4
types-python-dateutil     2.9.0.20250822
typing_extensions         4.15.0
typing-inspection         0.4.1
tzdata                    2025.2
umap-learn                0.5.9.post2
uri-template              1.3.0
urllib3                   2.5.0
wcwidth                   0.2.13
webcolors                 24.11.1
webencodings              0.5.1
websocket-client          1.8.0
websockets                15.0.1
